requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
aiofiles
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
python-jose
python-dotenv